Update contracts to Threshold Network version
NuCypher Network is going to merge with Keep Network, that will result
in the Threshold Network. This implies several changes to integrate the
NuCypher contracts with Threshold architecture.

In order to get this integration easier, NuCypher contracts have been
added to this repository, which will become the main place where the
contracts are stored.

This commit update the contracts to a version that enables the
integration with the Threshold Network. The contracts have been copied
from repo nucypher/nucypher branch threshold-network, commit b28b54adf.