Add return of error code (1) for initiate_ritual
Up to now, the script initiate_dkg.sh, aimed to be used on GitHub
workflows, doesn't return an error code even if the initiate_ritual ape
script fails. So the Github actions seems to be executed properly when
it is not.

With these changes, the Github workflow will appear as failed when the
script fails.