Fix node removal from excluded nodes list
When we add to the excluded nodes list a node that is not part of the
round (maybe because the node is shut down, so the script is not
considering it to be part of the heartbeat rond), the script returns an
error since it is trying to remove something (`node`) from a list that
is not containing that element.

This change fixes it.