Fix get_valid_versions function
The previous implementation calculates the deadline of each node version
based on the publication date of the same version. For example, for the
version 7.6.0, the release date is Aug 18, so the deadline would be Sep
8. This is not correct.

The deadline must be calculated based on the publication date of the
next release. So, for version 7.6.0, we have to take the publication
date of the next version 7.6.1: Sep 24. So the deadline is Oct 15.