Remove prettier pre-commit hook
Since we don't have javascript or typescript code here, and the
formatting and linting of Python code is made by black and flake8, it is
no need to check every commit with prettier.

Solidity linting tool is kept here. This uses prettier, but as long as
the npm package is installed, it will continue working when solidity
files are modified.